Flask App Implementation
Convert the entire app to use Flask App and Convert to Python 3.10

+++ website/models.py
@@ -9,7 +9,7 @@ class User(db.Model, UserMixin):
     name = db.Column(db.String(100),index=True,unique=True,nullable=False)
     emailid = db.Column(db.String(100), index=True, nullable=False)
     password_hash = db.Column(db.String(255),nullable=False)
-    contact_number = db.Column(db.Integer(10), index=True, nullable=False)
+    contact_number = db.Column(db.Integer, index=True, nullable=False)
     street_address = db.Column(db.String(100), index=True, nullable=False)
 
     comments = db.relationship('Comment',backref='user')
@@ -23,7 +23,7 @@ class Event(db.Model):
     description=db.Column(db.String(1000), nullable=False)
     location=db.Column(db.String(200),nullable=False)
     time=db.Column(db.DateTime,nullable=False)
-    contact_number = db.Column(db.Integer(10), index=True, nullable=False)
+    contact_number = db.Column(db.Integer, index=True, nullable=False)
     street_address = db.Column(db.String(100), index=True, nullable=False)
 
     comments = db.relationship('Comment',backref='events')
